feat(regex-actions): add new backend system support
