+++ plugins/regex-actions/dist-dynamic/package.json
@@ -13,6 +13,17 @@
   "backstage": {
     "role": "backend-plugin-module"
   },
+  "exports": {
+    ".": {
+      "require": "./dist/index.cjs.js",
+      "default": "./dist/index.cjs.js"
+    },
+    "./alpha": {
+      "require": "./dist/alpha.cjs.js",
+      "default": "./dist/alpha.cjs.js"
+    },
+    "./package.json": "./package.json"
+  },
   "scripts": {
     "build": "backstage-cli package build",
     "clean": "backstage-cli package clean",
@@ -31,7 +42,8 @@
   },
   "devDependencies": {},
   "files": [
-    "dist"
+    "dist",
+    "alpha"
   ],
   "repository": "github:janus-idp/backstage-plugins",
   "keywords": [
