+++ plugins/regex-actions/src/alpha.ts
@@ -1,0 +1,1 @@
+export { scaffolderModuleRegexActions as default } from './module';
