+++ plugins/regex-actions/src/module.ts
@@ -1,0 +1,19 @@
+import { createBackendModule } from '@backstage/backend-plugin-api';
+import { scaffolderActionsExtensionPoint } from '@backstage/plugin-scaffolder-node/alpha';
+
+import { createReplaceAction } from './actions';
+
+export const scaffolderModuleRegexActions = createBackendModule({
+  moduleId: 'scaffolder-backend-regexp',
+  pluginId: 'scaffolder',
+  register(env) {
+    env.registerInit({
+      deps: {
+        scaffolder: scaffolderActionsExtensionPoint,
+      },
+      async init({ scaffolder }) {
+        scaffolder.addActions(createReplaceAction());
+      },
+    });
+  },
+});
